+++ packages/js/src/types.ts
@@ -65,6 +65,10 @@ export interface AddEcommerceItemParams {
   productQuantity?: number
 }
 
+export interface RemoveEcommerceItemParams {
+  sku: string
+}
+
 export interface SetEcommerceViewParams {
   sku: string | boolean
   productName?: string | boolean