+++ CHANGELOG.md
@@ -14,9 +14,9 @@ Prefix the change with one of these keywords:
 - _Fixed_: for any bug fixes.
 - _Security_: in case of vulnerabilities.
 
-## [Unreleased]
-
+## [0.5.0]
 - Added: Support outbound links on `<a>` child elements.
+- Added: `removeEcommerceItem` and `clearEcommerceCart` action
 
 ## [0.4.0]